requirements.txt
# === core ML ===
# transformers 4.40+ uses new accelerate unwrap_model() signature.
# accelerate 1.0+ is required to match.
torch>=2.1.0
transformers>=4.40.0,<5.0
accelerate>=1.0,<2.0
peft>=0.9.0                     # DoRA requires >= 0.9
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4

# === OPTIONAL: int8 quantization (GPU only) ===
# Uncomment to enable `dtype: int8` in configs.
# bitsandbytes>=0.43

# === data / metrics ===
datasets>=2.18.0
scikit-learn>=1.3.0
nltk>=3.8
rouge-score>=0.1.2
numpy>=1.24

# === schema + serving + config ===
pydantic>=2.6
fastapi>=0.110
uvicorn[standard]>=0.27
pyyaml>=6.0

# === web scraper tool ===
requests>=2.31
beautifulsoup4>=4.12

# === UI (optional — for demo/gradio in Colab) ===
gradio>=4.0

# === dev / test ===
pytest>=8.0
matplotlib>=3.8
